Adds alignment for inner blocks that support alignment controls
As per this comment https://github.com/WordPress/gutenberg/pull/13964#issuecomment-472912261 the Block needs the ability to align the blocks within the Section. Currently only a few Blocks support this but a wider change can be added that will enable all inner blocks to be aligned. The alignment rules are tested against the [Gutenberg Starter Theme](https://github.com/WordPress/gutenberg-starter-theme) as this enqueues no editor styles and conforms to the guidlines laid out [here](https://github.com/WordPress/gutenberg/pull/13964#issuecomment-472562800).

@kjell has noted that Twenty Nineteen will need a patch for this, the styles for which have already been scoped out.
Note that this removes the previous fix for the block mover controls which are now (again) hidden when Blocks are full width. @kjell has confirmed this is a known issue that needs to be solved globally and should not be addressed in this PR.